fix: delete redundant header file; fix bug of size_t



# message auto-generated for no-merge-commit merge:
!29 merge fix_branch into main

fix: delete redundant header file; fix bug of size_t

Created-by: m0_69551548
Commit-by: Tom
Merged-by: fly2x
Description: fix: delete redundant header file; fix bug of size_t

See merge request: openHiTLS/pqcp!29

+++ test/sdv/kem/mceliece_test.c
@@ -144,7 +144,7 @@ static void CommitCurrentBlock(int pkBytes, int skBytes, int ctBytes)
 }
 
 static void ObtainMcElieceBytes(
-    PQC_Mceliece_AlgWithParamId algID, size_t *pkBytes, size_t *skBytes, size_t *ctBytes)
+    PQC_Mceliece_AlgWithParamId algID, int *pkBytes, int *skBytes, int *ctBytes)
 {
     switch (algID)
     {
